/numbers - Remove unhelpful comparisons with earlier elections
These comparisons were introduced when we just had the 2010 and 2015
UK elections on the site. Unfortunately, the code doesn't work now that
there are also local election and by-elections in the database. It's
better to just remove this feature than to present bizarre comparisons
to users.  The implementation would need to be completely rethought,
and that's very low priority compared to the other things we're trying
to get done in time for the 2017 General Election.

This is an alternative to fixing the problems described in #106