Photo moderation: never ask for a rejection reason for automated imports
The photo moderation interface asked you to enter a reason for rejecting
an image, since if an image was uploaded by someone that we can't
approve for the site, it's only fair to email them the reason why.

However, now there is also a script that automatically imports Twitter
avatar images for candidates and adds them to the queue. In these cases
the uploading_user will be None and any reason for rejection is
discarded (since there is no address to email it to). So we shouldn't be
asking people to fill in a box a box with that reason pointlessly.

Thanks to @davidmiller for pointing out this problem.