Add a function to infer the tree structure of versions with merges
When adding merge functionality to the site, we took the decision to
pretend that the version history of a person was still linear. While
this was simpler to do at the time, this has turned out to be a bit
problematic, particularly when it comes to reading the version history
and figure out what's going on. As a basis for making more readable
history and potentially migrating to explicitly storing the parents of
each version, we need to be able to infer the tree structure from the
existing versions array. This commit adds a function that does that.