Improve the rendering of diffs in the version history
This commit improves the rendering of diffs by making sure that they
are calculated against each 'parent' version or, if it is an initial
version, calculated against an empty dictionary. (Before we calculated
the diff against whatever was the previous version in the artificial
linear ordering of versions, which produced really confusing diffs in
some cases.)

Note that this contains a lot of updates to the test data and
expectations in test_version_diffs - the test data previously mostly
had no timestamp, version_id or person ID, which are now required to
show merges correctly. In addition, some of the timestamps which were
there conflicted with the order of the versions - the timestamps
have now been changed to match.