Add version IDs (including parents) as HTML element attributes
The id element of the version <div>s is in part to make it possible to
link to a particular version in the history. This also adds the parent
version IDs as a data-parent-version-ids attributes - this may also help
with Javascript enhancement of the version history (e.g. as @sjorford
has done).