QueuedImage admin: don't include the massive 'Person' drop-down
Looking at the detailed admin page for a QueuedImage caused a massive
slowdown on the site because it would try to populate a drop-down with
every candidate on the site in it for the "Person" field. We'll never
need to change the person a queued image refers to in the admin, so it's
simplest just to remove this field entirely from the detail page.