Require a source_url for each OfficialDocument
The bulk_adding application has some surprising behaviour if you ever
try to add candidates from a document without a source_url, and in
practice there should always be a source_url for every document. (We've
made sure this is the case in the UK database now.)

This commit removes the blank=True from the source_url field definition
so creating an OfficialDocument with a blank source_url should be
disallowed in the upload form and the admin.