Fix for not being able to add people through bulk_adding
A stray comma meant that a tuple was being assigned to source_url
instead of the URL of the document, and validation of the page where you
enter everyone's name was failing. This commit fixes that and adds a
functional test for the simplest possible workflow where you actually
create a person from the bulk_adding interface.