Make ConstituencyDetailView use PostExtraElection rather than Post
This change makes ConstituencyDetailView use PostExtraElection
as it's main object that's used for get_object_or_404 and
takes the Post[Extra] objects from the PostExtraElection.

This is an initial change that adds `post_election` in to the
context. Future changes should use post_election wherever possible
to avoid having to guess the post/election pairing.