Show results feature on all elections, not just 'local'
The reason for removing elections without 'local' in the slug was
that 'local' elections were assumed to all use the FPTP voting system.

This isn't true!

Because of this, we may as well show results for all elections,
this has the advantage of enabling recording results for General
Elections.

We need to filter out non-FPTP elections in future, but that's
another task for another day.