Fix API requests that are missing a trailing slash
Since the change introduced in 15050c8c45c508da2e24c2c8bebcf868fb3b6781
we have been getting the following error:

  Internal Server Error: /api/v0.9/posts/WMC:E14001021
  Traceback (most recent call last):
    File "/var/www/ynr/env/local/lib/python2.7/site-packages/django/core/handlers/base.py", line 223, in get_response
      response = middleware_method(request, response)
    File "/var/www/ynr/code/candidates/middleware.py", line 97, in process_response
      if request.user.is_authenticated():
  AttributeError: 'WSGIRequest' object has no attribute 'user'

... for API requests that are missing the trailing slash. In these cases
a redirect is being generated, but user is not being set on the request
object before it's processed by the DisableCachingForAuthenticatedUsers
middleware.  This fix seems to fix that problem.