"Upload photo" overlay for candidates without photos
If a candidate has no photo, and the user is logged in, we place an
"Upload photo" overlay over the candidate’s (placeholder) profile
picture in the `person-view` page header.

This makes the `person-view` and `person-edit` pages a bit more
consistent – if a candidate has no photo, you can always click the
placeholder image to upload a photo, regardless of what page you’re on.