Tidy up /feeds/needs-review.xml output
This fixes the double-escaping of double quotes, and removes the
unnecessary "Updated" line - there's already an <updated> element that
the consuming software can use instead, and it's noise in the situation
we want it for (a feed to review in Slack).

This commit also makes the expected XML in the test more readable again.