Refactor repeated tests into a mixin to be used by two person-views
Since we updated both the candidates person view and the UK elections
person view, the two tests have exactly the same tests, hence this
extraction of the common tests into a mixin with the shared code.

Code that is common to both templates should be placed in the shared
code mixin, while code specific to each template can go in its
dedicated test file.