Extract electon guessing code to it's own property
This is a continuation of the existing hack that guesses the
PostExtraElection based on the post. As the existing comment suggests
this is prone to errors and should be changes.

The reason for this change is that previously the guesses
PostExtraElection wasn't exposed to templates, so it wasn't possible
to use the PostExtraElection IDs in URLs.