Inferring merge parents: cope with secondary's history being missing
When merges were done in some period before the deployment of b3b2cfb on
2015-08-04, the version history of the secondary person wasn't included.
This was causing errors when viewing the pages of people who had other
records merged into theirs in that period.  This commit updates the code
that infers the parents of each version to cope with this situation, by
treating "After merging person [...]" versions as normal versions rather
than merge versions.

This should fix all but one of the problems described in #194