Split the upload photo view into a GET view and a POST view
This separation will help when we add the second form which also
makes a POST request but the contents sent are treated differently.
Having this view handle both the GET and the POST request to this
page would make it tricky to handle the two forms being in the
same page. This separation will hopefully make it clearer what is
going on.