Separate out your own post lock suggestions and put them at the bottom
As suggested by Sam Smith (@samsmith), when you go to the page of
suggested post locks to review, you're not meant to be reviewing your
own suggestions, so this commit separates out your own suggestions and
puts them in a section at the bottom of the page.

This commit also makes being authenticated a requirement to view this
page - this isn't a big deal, since you need to be authenticated and
have the ability to lock to see the link to this page anyway. (Not
having to deal with anonymous users makes the logic slightly simpler.)

Fixes #99