Add a command to cache the ouput of /api/v0.9/persons/ to a directory
When a new instance of WCIVF (whocanivotefor.co.uk) is brought up, it goes
through every page of /api/v0.9/persons/ to get the candidate data. This
ties up one worker of the server for a long time in any case, but in order
to scale WCIVF we might need to bring up multiple instances nearly
simultaneously which would really make the YNR server suffer. To avoid
this impacting the YNR instance too much, we've decided to cache the
results of these calls to disk (once a day probably) where they can be
served effeciently as static files by nginx.  WCIVF can then do a cheaper
API query to find just the people who have changed since a particular time.

The new command introduced in this commit should make it easy to
generate a version of this data in a format that is easy to
serve as static files. You can invoke it, for example, like:

  ./manage.py candidates_cache_persons_api \
     /var/www/ynr-media/cached-persons \
     https://candidates.democracyclub.org.uk/media/cached-persons

That will take some minutes to run, but will create a directory
structure under /var/www/ynr-media that looks like:

└── cached-persons
    ├── 2017-05-14T17:03:01
    │   ├── page-000001.json
    │   ├── page-000002.json
    │   ├── page-000003.json
            ...
    │   └── page-000165.json
    ├── 2017-05-14T17:09:13
    │   ├── page-000001.json
    │   ├── page-000002.json
    │   ├── page-000003.json
            ...
    │   └── page-000165.json
    └── latest -> 2017-05-14T17:09:13

... which you should then be able to iterate over as with the API as
normal, starting at:

  https://candidates.democracyclub.org.uk/media/cached-persons/latest/page-000001.json

... since the command rewrites the "next" and "previous" URLs in the
response to go to the correct page-XXXXXX.json file's URL. This means
that nginx will have to be configured to follow symlinks when serving
static files.

This uses the Django test client to fetch the API results, which seems
to be a pretty simple approach.

(This was @symroe's idea - I just did the implementation.)