Reuse database connections for up to a minute when DEBUG is False
This commits sets CONN_MAX_AGE in the DATABASE setting to 60 seconds,
meaning that a database connection can be reused for up to a minute by a
thread. We hope this will improve performance since many queries are
used to produce some pages, and reconnecting to the database server each
time is a waste.

The documentation for this feature, with a couple of important caveats,
is here:

   https://docs.djangoproject.com/en/1.8/ref/databases/#persistent-connections

Note the comment I've added, referring to the problem that this is
likely to be ineffective (or harmful) with eventlet or gevent workers.