Don't unnecessarily delete and recreate memberships
This affects the code called by UpdatePersonView and BulkAddReviewView
which previously deleted all memberships and created new ones when
updating someone. Unfortunately, since the addition of the uk_results
app, the CandidateResult model has a ForeignKey that points to
Membership, which meant that whenever a person was updated, any
CandidateResult that referred to them would be deleted.

This changes the behaviour of those two views to preserve an existing
candidacy membership if possible, which should mean that the result data
isn't lost via those code paths.

Note that there is another function that might still lose results data:
revert_person_from_version_data, which is called when reverting and
merging. When merging, there is a special case code to preserve any
CandidateResult, but it's still possible to lose results when reverting.
A subsequent commit will make sure an exception is thrown rather that
losing data when reverting.