Cache posts data as well as persons to the filesystem
The candidates_cache_persons_api command, which caches API results from
/api/v0.9/persons/ to the filesystem so they can be more efficiently
consumed by WCIVF, was limited to generating output from the persons
endpoint. It turns out that getting all the posts is also very
resource-consuming, however. This commit updates the command so that it
also outputs posts data. There are two changes to be particularly aware
of here:

 * The command has been renamed to candidates_cache_api_to_directory
   since it's not now limited to outputting data from /api/v0.9/persons.

 * Instead of files being called page-NNNNNNN.json, they're called
   persons-NNNNNN.json and posts-NNNNNN.json etc.