Add redirects for merged people to the CSV output
This enables consumers of the CSV files to tell, when a person's row
disappears from that file, which other person they had been merged into.

This implementation is probably more complex than you'd think it should be
to avoid doing more than one query on PersonRedirect when generating a
CSV file.

Fixes #242