Apply a size limit to images added by URL
It would be bad if someone could bring down the site by pasting in an
URL to a multiple terabyte image, say, so this commit limits the size of
images that might be download.

It's still possible that if the server is (maliciously or accidentally)
very slow, this could tie up workers until the site no longer works. We
could get around this by enqueuing the download of these submitted URLs
and dealing them with celery, say, but the current implementation will
hopefully do for the moment.