Fix the Twitter username updater sometimes losing version data
The Twitter username updater command would iterate over all people (with
a select_related on 'extra' to bring in the corresponding PersonExtra
data) and then open a transaction to handle each of those people. That's
a bug, since the loop takes a long time and if there's been an update
to a person since the loop started but before that person is reached by
the loop and the update then makes any changes to their Twitter data,
it'll save versions data that doesn't include the manual update that
had been made. (This has actually happened.)

Doing all these updates in a single transaction would mean having a
transaction open for a long time, which could cause other updates to
block, so this commit changes it to use a transaction per person, at the
cost of more queries to the database being needed.

Fixes #262