Cope with the version in a LoggedAction not being found for its person
There seem to be quite a few cases where a LoggedAction refers to a
version_id which isn't present in the person's version history. This
could happen due to the (now fixed) bad transaction handling in the
Twitter username update (#262), for example. At the moment this makes
/feeds/needs-review.xml return a 500 error - the presence of some of
these cases shouldn't stop the whole feed from working, so this commit
changes the behaviour to catch the exception if a version is not found
and output a warning in the HTML instead.