Only look for the last day's changes in the 'needs review' feed
With production data, the checks for potentially suspect changes take a
long time, and we've been getting bad gateway errors which I suspect are
due to the request timing out. Hopefully only considering a fifth as
much data (1 day instead of 5), as this commit does, will at least get
it working again to some extent.