Don't pretend we know IEBC candidate numbers
In the candidate data from the original PDF, some rows were missing a
number in the 'No' column. Originally I filled in some monotonically
increasing numbers, but on reconsideration we shouldn't do that in cases
except when there's very clearly just a missing number in the sequence.

This commit changes the more dubious suggested identifiers for some
prefixed with 'unknown:'.