Cope with get_current_usersettings returning an EmptyUserSettings
In some important situations (running tests and creating a site from
scratch), no SiteSettings object will exist for this site, so
get_current_usersettings returns EmptySiteSettings, and any attempts
to access any of the settings fields will raise an exception.

This commit adds a get_site_setting function that will provide return
the default value for the field (or None if there was no default) in the
case that the attribute isn't found on the object returned from
get_current_usersettings.