Merge two independent branches of migrations
The DemocracyClub and mySociety forks of YNR now have a different
migration history in the candidates app. At this point we have
backported some of the code + migrations from the DemocracyClub fork,
but not all. This means that the code that appplies migrations when
creating a dependency graph can't tell which is the most recent - this
commit adds a dependency from one of the mySociety fork's migrations
to one of those from the DemocracyClub fork, so that Django migrations
can order these migrations.

Up to this point, the migrations in the two forks are genuinely
independent, I believe, so there should be no problem with conflicts.
There are further migrations to come, so at some point we'll have to
make a similar commit to merge the later migration history as well.